Add files via upload
we all know that we have many languages across the world and it is inconceivable to learn all those languages. here language translation plays a crucial role and there is a need of language translator. So I came up with a project called "language translator and text to speech convertor" in aws cloud. 
In order to accomplish this project, I used EC2, IAM, Amazon translate and Amazon polly services. 
Amazon translate is neutral machine translation service and it uses deep learning techniques to provide fast , accurate and high quality translation. Amazon polly service converts text into lifelike speech. Using of these two services made our translation and conversion  easy. 
In this project video, you are able to see the translation of give source text  to required target text and able to hear the pronunciation of text. 
This project solves the problem raised about language and helps in spreading new information, knowledge and ideas across the world. It also enhances the international marketing and expands the business.